+++ c++/pod5_format/file_writer.cpp
@@ -625,11 +625,52 @@ pod5::Result<std::unique_ptr<FileWriter>> create_file_writer(
         pool));
 }
 
+static Status add_recovery_failure_context(
+    Status status,
+    std::string const & tmp_path,
+    std::string const & description)
+{
+    assert(!status.ok());
+    std::string const error_context =
+        "Failed whilst attempting to recover " + description + " from file - " + tmp_path;
+    if (status.detail()) {
+        return status.WithMessage(error_context);
+    }
+    return arrow::Status::FromArgs(status.code(), error_context + ". Detail: " + status.message());
+}
+
+template <typename writer_type>
+static Status append_recovered_file(
+    std::string const & tmp_path,
+    writer_type const & destination_writer,
+    std::string const & description,
+    arrow::MemoryPool * const pool)
+{
+    arrow::Status inner_status = [&] {
+        ARROW_ASSIGN_OR_RAISE(auto file, arrow::io::ReadableFile::Open(tmp_path, pool));
+        ARROW_ASSIGN_OR_RAISE(auto size, file->GetSize());
+        if (size == 0) {
+            return arrow::Status::Invalid("File is empty/zero bytes long.");
+        }
+        ARROW_ASSIGN_OR_RAISE(
+            RecoveredData const recovered_raw_data, recover_arrow_file(file, destination_writer));
+        return arrow::Status::OK();
+    }();
+    if (!inner_status.ok()) {
+        return add_recovery_failure_context(inner_status, tmp_path, description);
+    }
+    return inner_status;
+}
+
 pod5::Result<std::unique_ptr<FileWriter>> recover_file_writer(
     std::string const & src_path,
     std::string const & dest_path,
     FileWriterOptions const & options)
 {
+    if (!check_extension_types_registered()) {
+        return arrow::Status::Invalid("POD5 library is not correctly initialised.");
+    }
+
     // Create a file to push recovered data into:
     ARROW_ASSIGN_OR_RAISE(
         auto dest_file, create_file_writer(dest_path, "pod5_file_recovery", options));
@@ -642,44 +683,32 @@ pod5::Result<std::unique_ptr<FileWriter>> recover_file_writer(
     // Signature should be right at 0:
     ARROW_RETURN_NOT_OK(combined_file_utils::check_signature(file, 0));
 
-    auto null_metadata = arrow::KeyValueMetadata::Make({}, {});
-
     // Recover the signal data into [dest_file]:
-    RecoveredData recovered_raw_data;
+    arrow::Result<RecoveredData> recovered_raw_data;
     {
         ARROW_ASSIGN_OR_RAISE(
             auto raw_sub_file,
             combined_file_utils::open_sub_file(file, combined_file_utils::header_size));
-        ARROW_ASSIGN_OR_RAISE(
-            recovered_raw_data,
-            recover_arrow_file(raw_sub_file, dest_file->impl()->signal_table_writer()));
+        recovered_raw_data =
+            recover_arrow_file(raw_sub_file, dest_file->impl()->signal_table_writer());
+    }
+    if (!recovered_raw_data.ok()) {
+        return add_recovery_failure_context(
+            recovered_raw_data.status(), arrow_path.ToString(), "signal data sub file");
     }
 
-    auto file_identifier = recovered_raw_data.metadata.file_identifier;
-    auto reads_tmp_path = make_reads_tmp_path(arrow_path, file_identifier);
-    auto run_info_tmp_path = make_run_info_tmp_path(arrow_path, file_identifier);
+    auto file_identifier = recovered_raw_data->metadata.file_identifier;
 
     // Recover the run info data into [dest_file]:
-    {
-        ARROW_ASSIGN_OR_RAISE(auto file, arrow::io::ReadableFile::Open(run_info_tmp_path, pool));
-        ARROW_ASSIGN_OR_RAISE(auto size, file->GetSize());
-        if (size > 0) {
-            ARROW_ASSIGN_OR_RAISE(
-                recovered_raw_data,
-                recover_arrow_file(file, dest_file->impl()->run_info_table_writer()));
-        }
-    }
+    auto run_info_tmp_path = make_run_info_tmp_path(arrow_path, file_identifier);
+    auto run_info_writer = dest_file->impl()->run_info_table_writer();
+    ARROW_RETURN_NOT_OK(
+        append_recovered_file(run_info_tmp_path, run_info_writer, "run information", pool));
 
     // Recover the read data into [dest_file]:
-    {
-        ARROW_ASSIGN_OR_RAISE(auto file, arrow::io::ReadableFile::Open(reads_tmp_path, pool));
-        ARROW_ASSIGN_OR_RAISE(auto size, file->GetSize());
-        if (size > 0) {
-            ARROW_ASSIGN_OR_RAISE(
-                recovered_raw_data,
-                recover_arrow_file(file, dest_file->impl()->read_table_writer()));
-        }
-    }
+    auto reads_tmp_path = make_reads_tmp_path(arrow_path, file_identifier);
+    auto read_writer = dest_file->impl()->read_table_writer();
+    ARROW_RETURN_NOT_OK(append_recovered_file(reads_tmp_path, read_writer, "reads", pool));
 
     return dest_file;
 }